reorganize/generalize tensorboard callback
